feat(KP-332): SearchDepartment is now converted to TS
