+++ public/js/app/components/SearchDepartments/types.ts
@@ -1,0 +1,6 @@
+export { SchoolsWithDepartments } from '../../stores/types/searchPageStoreTypes'
+
+export interface SearchDepartmentsProps {
+  onChange: (selected: { department: string }) => void
+  disabled?: boolean
+}
